+++ tests/test_pgml_lint_parser.py
@@ -11,7 +11,7 @@ def test_newline_index_and_pos_to_line() -> None:
 	assert pgml_lint.parser.pos_to_line(newlines, 0) == 1
 	assert pgml_lint.parser.pos_to_line(newlines, 2) == 2
 	assert pgml_lint.parser.pos_to_line(newlines, 4) == 3
-	assert pgml_lint.parser.pos_to_line(newlines, 5) == 4
+	assert pgml_lint.parser.pos_to_line(newlines, 5) == 3
 
 
 #============================================